+++ crates/create-farm-rs/templates/electron/vue/farm.config.ts
@@ -1,4 +1,4 @@
-import { defineConfig } from '@farmfe/core';
+import { defineConfig } from 'farm';
 import electron from '@farmfe/js-plugin-electron';
 import vue from '@vitejs/plugin-vue';