+++ js-plugins/electron/farm.config.mjs
@@ -1,16 +1,16 @@
-import farmDtsPlugin from '@farmfe/js-plugin-dts';
-import { createFarmJsPluginBuildConfig } from '../../configs/farm-js-plugin.base.config.mjs';
+import farmDtsPlugin from "@farmfe/js-plugin-dts";
+import { createFarmJsPluginBuildConfig } from "../../configs/farm-js-plugin.base.config.mjs";
 
 /**
- * @type {import('@farmfe/core').UserConfig}
+ * @type {import('farm').UserConfig}
  */
 export default createFarmJsPluginBuildConfig(
   [
     farmDtsPlugin({
-      tsConfigPath: './tsconfig.build.json'
-    })
+      tsConfigPath: "./tsconfig.build.json",
+    }),
   ],
   {
-    external: ['electron'],
+    external: ["electron"],
   },
 );