+++ packages/create-farm-plugin/CHANGELOG.md
@@ -113,7 +113,7 @@
 ### Patch Changes
 
 - 3a51e6c0: chore(deps-dev): bump @farmfe/cli from 0.6.2 to 1.0.0 in /packages/create-farm-plugin/templates/js
-- d54eacc3: chore(deps-dev): bump @farmfe/core from 0.16.10 to 1.0.5 in /packages/create-farm-plugin/templates/js
+- d54eacc3: chore(deps-dev): bump farm from 0.16.10 to 1.0.5 in /packages/create-farm-plugin/templates/js
 
 ## 0.1.1